chore: extract timeline data into a timeline page
